pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "ai-financial-advisor"
version = "0.2.0"
description = "A data-driven, AI-powered investment advisory system that integrates news analysis, technical indicators, and market sentiment to provide actionable investment insights."
readme = "README.md"
license = "MIT"
requires-python = ">=3.11"
authors = [
    { name = "Jeff Liu" },
]
keywords = ["finance", "ai", "investment", "nlp", "quantitative-trading"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "Intended Audience :: Financial and Insurance Industry",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
    "Topic :: Office/Business :: Financial :: Investment",
]

dependencies = [
    "pydantic-settings>=2.0",
    "openai>=1.0",
    "anthropic>=0.30",
    "newsapi-python>=0.2.7",
    "newspaper3k>=0.2.8",
    "lxml-html-clean>=0.4",
    "yfinance>=0.2",
    "pandas>=2.0",
    "numpy>=1.24",
    "plotly>=5.0",
    "jinja2>=3.1",
    "typer[all]>=0.12",
    "fredapi>=0.5",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "pytest-cov>=5.0",
    "ruff>=0.5",
    "mypy>=1.10",
    "pre-commit>=3.0",
    "httpx",
]
web = [
    "gradio>=4.0",
    "fastapi>=0.110",
    "uvicorn>=0.29",
]

[project.scripts]
ai-advisor = "ai_financial_advisor.cli:app"

[tool.hatch.build.targets.wheel]
packages = ["src/ai_financial_advisor"]

[tool.ruff]
target-version = "py311"
src = ["src", "tests"]
line-length = 120

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W", "UP"]

[tool.mypy]
python_version = "3.11"
strict = true

[tool.pytest.ini_options]
testpaths = ["tests"]
addopts = "-v --tb=short"
